cover chat methods

+++ telegram-bot/src/main/kotlin/eu/vendeli/tgbot/types/internal/options/ChatInviteLinkOptions.kt
@@ -2,7 +2,7 @@ package eu.vendeli.tgbot.types.internal.options
 
 data class ChatInviteLinkOptions(
     var name: String? = null,
-    var expireDate: Int? = null,
+    var expireDate: Long? = null,
     var memberLimit: Int? = null,
     var createsJoinRequest: Boolean? = null,
 ) : Options
